Basic Capital Spawning Added to Inits
Note to Future Puck - System Ownership, Buildings, Deposits, Starbases, Colonies etc still need to all be done. Also double check the scripted effects you're calling on some of the inits, as not all empires are spawning in, so when you're back to work on things, check the missing empires first. Ignore the 0 pop subspecies, you've wasted a week trying to track it down and it only brought you pain and misery.