Hanar,Batarian,Krogan,Omega Spec Sheet Review
Should be as much of the spec sheet review done as possible. Some things like the Batarian Raiding Stance must be done in the raiding stance file which we haven't copied over. Also some things like happiness when at war seems to not be doable eventhough i swear i've seen it before.